+++ tests/NeuralNet/Initializers/ConstantTest.php
@@ -1,0 +1,39 @@
+<?php
+
+namespace Rubix\ML\Tests\NeuralNet\Initializers;
+
+use Rubix\Tensor\Matrix;
+use Rubix\ML\NeuralNet\Initializers\Constant;
+use Rubix\ML\NeuralNet\Initializers\Initializer;
+use PHPUnit\Framework\TestCase;
+
+class ConstantTest extends TestCase
+{
+    protected $initializer;
+
+    public function setUp()
+    {
+        $this->initializer = new Constant(4.8);
+    }
+
+    public function test_build_layer()
+    {
+        $this->assertInstanceOf(Constant::class, $this->initializer);
+        $this->assertInstanceOf(Initializer::class, $this->initializer);
+    }
+
+    public function test_initialize()
+    {
+        $w = $this->initializer->initialize(4, 3);
+
+        $expected = [
+            [4.8, 4.8, 4.8, 4.8],
+            [4.8, 4.8, 4.8, 4.8],
+            [4.8, 4.8, 4.8, 4.8],
+        ];
+
+        $this->assertInstanceOf(Matrix::class, $w);
+        $this->assertEquals([3, 4], $w->shape());
+        $this->assertEquals($expected, $w->asArray());
+    }
+}